cmd/gc: do not nop-convert equivalent but different interface types.
The cached computed interface tables are indexed by the interface
types, not by the unnamed underlying interfaces

To preserve the invariants expected by interface comparison, an
itab generated for an interface type must not be used for a value
of a different interface type even if the representation is identical.

Fixes issue 7207.

LGTM=rsc
R=rsc, iant, khr
CC=golang-codereviews
https://codereview.appspot.com/69210044